+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Added footer extension points for links and avatar
+- Create OpenAPI specification during build
 - Extension point entries with supplied extensionName are sorted ascending
+
+### Changed
+- New footer design
 
 ### Fixed
 - Modification for mercurial repositories with enabled XSRF protection
+
+### Removed
+- Enunciate rest documentation
 
 ## 2.0.0-rc4 - 2020-02-14
 ### Added